@@ -3,36 +3,38 @@ authors = ["Alexander Sidorov <alexander@sidorov.dev>"]
 description = "Example: Django + MSSQL + Docker"
 license = "MIT"
 name = "example_django_mssql_docker"
-version = "2023.06.01"
+version = "2023.08.01"
 
 
 [tool.poetry.dependencies]
-devtools = {extras = ["pygments"], version = "^0.11.0"}
-dj-database-url = "^2.0.0"
-django = "^4.2.2"
-djangorestframework = "^3.14.0"
-gunicorn = "^20.1.0"
-mssql-django = "^1.3"
-psycopg2-binary = "^2.9.6"
-pydantic = {extras = ["dotenv"], version = "^1.10.9"}
-pyodbc = "^4.0.39"
+devtools = {extras = ["pygments"], version = "0.11.0"}
+dj-database-url = "2.0.0"
+django = "4.2.3"
+django-mssql-backend = "2.8.1"
+djangorestframework = "3.14.0"
+gunicorn = "21.2.0"
+mssql-django = "1.3"
+psycopg2-binary = "2.9.6"
+pydantic = "2.1.0"
+pydantic-settings = "2.0.2"
+pyodbc = "4.0.39"
 python = "3.11.3"
-requests = "^2.31.0"
-whitenoise = {extras = ["brotli"], version = "^6.5.0"}
+requests = "2.31.0"
+whitenoise = {extras = ["brotli"], version = "6.5.0"}
 
 [tool.poetry.group.dev.dependencies]
 autoflake = "^2.2.0"
-black = "^23.3.0"
-django-stubs = "^4.2.2"
+black = "^23.7.0"
+django-stubs = "^4.2.3"
 djangorestframework-stubs = "^3.14.2"
 flake8 = "^6.0.0"
-flake8-annotations-complexity = "^0.0.7"
+flake8-annotations-complexity = "^0.0.8"
 flake8-bandit = "^4.1.1"
 flake8-black = "^0.3.6"
-flake8-bugbear = "^23.6.5"
+flake8-bugbear = "^23.7.10"
 flake8-builtins = "^2.1.0"
 flake8-cognitive-complexity = "^0.1.0"
-flake8-comprehensions = "^3.13.0"
+flake8-comprehensions = "^3.14.0"
 flake8-debugger = "^4.1.2"
 flake8-executable = "^2.1.3"
 flake8-expression-complexity = "^0.0.11"
@@ -44,19 +46,19 @@ flake8-pep3101 = "^2.0.0"
 flake8-print = "^5.0.0"
 flake8-printf-formatting = "^1.1.2"
 flake8-simplify = "^0.20.0"
-flake8-variables-names = "^0.0.5"
+flake8-variables-names = "^0.0.6"
 httpie = "^3.2.2"
 ipython = "^8.14.0"
 isort = "^5.12.0"
 mypy = "^1.4.1"
 pep8-naming = "^0.13.3"
 pytest = "^7.4.0"
-types-psycopg2 = "^2.9.21.10"
+types-psycopg2 = "^2.9.21.11"
 
 
 [tool.black]
 line-length = 79
-target-version = ["py310"]
+target-version = ["py311"]
 include = '^.+\.pyi?$'
 extend-exclude = '''
 (
@@ -103,7 +105,7 @@ plugins = [
     "pydantic.mypy",
 ]
 pretty = true
-python_version = "3.10"
+python_version = "3.11"
 show_error_context = true
 sqlite_cache = true
 warn_redundant_casts = true
