@@ -1,15 +1,18 @@
-from pydantic import BaseSettings
+from pydantic_settings import BaseSettings
+from pydantic_settings import SettingsConfigDict
 
 from domains.fs import dirs
 
 
 class Config(BaseSettings):
-    class Config:
-        case_sensitive = True
-        env_file = dirs.DIR_REPO / ".env"
-        env_file_encoding = "utf-8"
-        env_prefix = "WEBAPP_"
-        frozen = True
+    model_config = SettingsConfigDict(
+        case_sensitive=True,
+        env_file_encoding="utf-8",
+        env_file=dirs.DIR_REPO / ".env",
+        env_prefix="WEBAPP_",
+        extra="ignore",
+        frozen=True,
+    )
 
     MODE_DEBUG: bool = False
     MSSQL_DATABASE_HOST: str
